requirements.txt
# Oil Trading Dashboard - Core Dependencies
# ==================================================

# Web Framework
streamlit>=1.28.0
streamlit-aggrid>=0.3.4

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
polars>=0.19.0
pyarrow>=14.0.0

# Database
duckdb>=0.9.0
sqlalchemy>=2.0.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
wordcloud>=1.9.0

# Caching & Storage
diskcache>=5.6.0
python-dateutil>=2.8.2
pytz>=2023.3

# API & Networking
requests>=2.31.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Background Tasks
apscheduler>=3.10.0

# Utilities
joblib>=1.3.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Financial Calculations
scipy>=1.11.0
